+++ .pre-commit-config.yaml
@@ -1,9 +1,4 @@
 repos:
--   repo: https://github.com/asottile/reorder_python_imports
-    rev: v1.4.0
-    hooks:
-    -   id: reorder-python-imports
-        types: [python]
 -   repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v2.2.1
     hooks:
@@ -13,6 +8,22 @@ repos:
     -   id: pretty-format-json
         args: [--autofix, --no-sort-keys]
     -   id: debug-statements
+-   repo: https://github.com/asottile/reorder_python_imports
+    rev: v1.4.0
+    hooks:
+    -   id: reorder-python-imports
+        types: [python]
+-   repo: https://github.com/python/black
+    rev: 19.3b0
+    hooks:
+    -   id: black
+        types: [python]
+-   repo: https://github.com/asottile/blacken-docs
+    rev: v0.5.0
+    hooks:
+    -   id: blacken-docs
+        additional_dependencies: [black]
+        types: [rst]
 -   repo: https://gitlab.com/pycqa/flake8
     rev: 3.7.7
     hooks:
@@ -24,17 +35,6 @@ repos:
             pep8-naming,
         ]
         files: 'respy/.*'  # We start with a subfolder and extend the scope later.
--   repo: https://github.com/asottile/blacken-docs
-    rev: v0.5.0
-    hooks:
-    -   id: blacken-docs
-        additional_dependencies: [black]
-        types: [rst]
--   repo: https://github.com/python/black
-    rev: 19.3b0
-    hooks:
-    -   id: black
-        types: [python]
 -   repo: local
     hooks:
     -   id: doc8
