+++ environment.yml
@@ -9,7 +9,6 @@ dependencies:
   - pandas=0.24.2
   - nbsphinx=0.4.2
   - numba=0.43.1
-  - scipy=1.2.1
   - codecov
   - doc8
   - jupyterlab
@@ -23,3 +22,5 @@ dependencies:
   - sphinx-autobuild
   - sphinx_rtd_theme
   - tox-conda
+  - pip:
+    - pre-commit
