+++ respy/config.py
@@ -5,21 +5,17 @@ import numpy as np
 
 # Obtain the root directory of the package. Do not import respy which creates a circular
 # import.
-ROOT_DIR = Path(__file__).parents[2]
+ROOT_DIR = Path(__file__).parent
 
 # Directory with additional resources for the testing harness
-TEST_DIR = ROOT_DIR / "tests"
 TEST_RESOURCES_DIR = ROOT_DIR / "tests" / "resources"
 
-MINISCULE_FLOAT = 1.0e-100
-LARGE_FLOAT = 1.0e8
 HUGE_FLOAT = 1.0e20
-SMALL_FLOAT = 1e-5
 TINY_FLOAT = 1.0e-8
 PRINT_FLOAT = 1e10
 
-# Number of decimals that are compared for tests
-# This is currently only used in regression tests.
+# Number of decimals that are compared for tests This is currently only used in
+# regression tests.
 DECIMALS = 6
 # Some assert fucntions take rtol instead of decimals
 TOL = 10 ** -DECIMALS
@@ -27,15 +23,7 @@ TOL = 10 ** -DECIMALS
 # Interpolation
 INADMISSIBILITY_PENALTY = -400000.00
 
-# Missing values. These allow to aline the treatment of missing values across
-# implementations. There is no NAN available in FORTRAN.
-MISSING_INT = -99
-MISSING_FLOAT = -99.00
-
 IS_DEBUG = False
-
-# Each implementation has its own set of optimizers available.
-OPTIMIZERS = ["SCIPY-BFGS", "SCIPY-POWELL", "SCIPY-LBFGSB"]
 
 # Labels for columns in a dataset as well as the formatters.
 DATA_LABELS_EST = [
@@ -74,11 +62,13 @@ DATA_LABELS_SIM = DATA_LABELS_EST + [
     "Immediate_Reward_4",
 ]
 
-DATA_FORMATS_EST = {}
-for key_ in DATA_LABELS_EST:
-    DATA_FORMATS_EST[key_] = np.int
-    if key_ in ["Wage"]:
-        DATA_FORMATS_EST[key_] = np.float
+DATA_FORMATS_EST = {
+    col: (np.float if col == "Wage" else np.int) for col in DATA_LABELS_EST
+}
+DATA_FORMATS_SIM = {
+    col: (np.int if col == "Type" else np.float) for col in DATA_LABELS_SIM
+}
+DATA_FORMATS_SIM = {**DATA_FORMATS_SIM, **DATA_FORMATS_EST}
 
 DATA_FORMATS_SIM = dict(DATA_FORMATS_EST)
 for key_ in DATA_LABELS_SIM:
