--- respy/python/interface.py
@@ -1,526 +1,0 @@
-from functools import partial
-from math import floor
-from math import log10
-
-import numpy as np
-from scipy.optimize import approx_fprime
-from scipy.optimize import fmin_bfgs
-from scipy.optimize import fmin_l_bfgs_b
-from scipy.optimize import fmin_powell
-
-from respy.custom_exceptions import MaxfunError
-from respy.python.estimate.estimate_python import pyth_criterion
-from respy.python.estimate.estimate_wrapper import OptimizationClass
-from respy.python.record.record_estimation import record_estimation_final
-from respy.python.record.record_estimation import record_estimation_scaling
-from respy.python.record.record_estimation import record_estimation_stop
-from respy.python.shared.shared_auxiliary import apply_scaling
-from respy.python.shared.shared_auxiliary import create_draws
-from respy.python.shared.shared_auxiliary import dist_class_attributes
-from respy.python.shared.shared_auxiliary import get_optim_paras
-from respy.python.shared.shared_constants import DATA_FORMATS_EST
-from respy.python.shared.shared_constants import DATA_LABELS_EST
-from respy.python.shared.shared_constants import HUGE_FLOAT
-from respy.python.simulate.simulate_python import pyth_simulate
-from respy.python.solve.solve_auxiliary import StateSpace
-from respy.python.solve.solve_python import pyth_solve
-
-
-def respy_interface(respy_obj, request, data=None):
-    """Provide the interface to the PYTHON functionality."""
-    # Distribute class attributes
-    (
-        optim_paras,
-        num_periods,
-        edu_spec,
-        is_debug,
-        num_draws_prob,
-        seed_prob,
-        num_draws_emax,
-        seed_emax,
-        interpolation,
-        num_points_interp,
-        maxfun,
-        optimizer_used,
-        tau,
-        optimizer_options,
-        seed_sim,
-        num_agents_sim,
-        file_sim,
-        precond_spec,
-        num_types,
-        num_paras,
-        num_agents_est,
-    ) = dist_class_attributes(
-        respy_obj,
-        "optim_paras",
-        "num_periods",
-        "edu_spec",
-        "is_debug",
-        "num_draws_prob",
-        "seed_prob",
-        "num_draws_emax",
-        "seed_emax",
-        "interpolation",
-        "num_points_interp",
-        "maxfun",
-        "optimizer_used",
-        "tau",
-        "optimizer_options",
-        "seed_sim",
-        "num_agents_sim",
-        "file_sim",
-        "precond_spec",
-        "num_types",
-        "num_paras",
-        "num_agents_est",
-    )
-
-    if request == "estimate":
-
-        periods_draws_prob = create_draws(
-            num_periods, num_draws_prob, seed_prob, is_debug
-        )
-        periods_draws_emax = create_draws(
-            num_periods, num_draws_emax, seed_emax, is_debug
-        )
-
-        # Construct starting values
-        x_optim_free_unscaled_start = get_optim_paras(
-            optim_paras, num_paras, "free", is_debug
-        )
-        x_optim_all_unscaled_start = get_optim_paras(
-            optim_paras, num_paras, "all", is_debug
-        )
-
-        # Construct the state space
-        state_space = StateSpace(
-            num_periods, num_types, edu_spec["start"], edu_spec["max"]
-        )
-
-        # Collect arguments that are required for the criterion function.
-        # These must be in the correct order already.
-        args = (
-            interpolation,
-            num_points_interp,
-            is_debug,
-            data,
-            tau,
-            periods_draws_emax,
-            periods_draws_prob,
-            state_space,
-        )
-
-        # Special case where just one evaluation at the starting values is
-        # requested is accounted for. Note, that the relevant value of the
-        # criterion function is always the one indicated by the class attribute
-        # and not the value returned by the optimization algorithm.
-        num_free = optim_paras["paras_fixed"].count(False)
-
-        # Take only bounds from unfixed parameters and insert default bounds.
-        mask_paras_fixed = np.array(optim_paras["paras_fixed"])
-        paras_bounds_free_unscaled = np.array(optim_paras["paras_bounds"])[
-            ~mask_paras_fixed
-        ]
-        paras_bounds_free_unscaled[:, 0] = np.where(
-            paras_bounds_free_unscaled[:, 0] == None,  # noqa: E711
-            -HUGE_FLOAT,
-            paras_bounds_free_unscaled[:, 0],
-        )
-        paras_bounds_free_unscaled[:, 1] = np.where(
-            paras_bounds_free_unscaled[:, 1] == None,  # noqa: E711
-            HUGE_FLOAT,
-            paras_bounds_free_unscaled[:, 1],
-        )
-
-        record_estimation_scaling(
-            x_optim_free_unscaled_start,
-            None,
-            None,
-            None,
-            optim_paras["paras_fixed"],
-            True,
-        )
-
-        precond_matrix = get_precondition_matrix(
-            precond_spec,
-            optim_paras,
-            x_optim_all_unscaled_start,
-            args,
-            maxfun,
-            num_paras,
-            num_types,
-        )
-
-        x_optim_free_scaled_start = apply_scaling(
-            x_optim_free_unscaled_start, precond_matrix, "do"
-        )
-
-        paras_bounds_free_scaled = np.full((num_free, 2), np.nan)
-        for i in range(2):
-            paras_bounds_free_scaled[:, i] = apply_scaling(
-                paras_bounds_free_unscaled[:, i], precond_matrix, "do"
-            )
-
-        record_estimation_scaling(
-            x_optim_free_unscaled_start,
-            x_optim_free_scaled_start,
-            paras_bounds_free_scaled,
-            precond_matrix,
-            optim_paras["paras_fixed"],
-            False,
-        )
-
-        opt_obj = OptimizationClass(
-            x_optim_all_unscaled_start,
-            optim_paras["paras_fixed"],
-            precond_matrix,
-            num_types,
-        )
-        opt_obj.maxfun = maxfun
-
-        if maxfun == 0:
-
-            opt_obj.crit_func(x_optim_free_scaled_start, *args)
-
-            success = True
-            message = "Single evaluation of criterion function at starting values."
-
-        elif optimizer_used == "SCIPY-BFGS":
-
-            bfgs_maxiter = optimizer_options["SCIPY-BFGS"]["maxiter"]
-            bfgs_gtol = optimizer_options["SCIPY-BFGS"]["gtol"]
-            bfgs_eps = optimizer_options["SCIPY-BFGS"]["eps"]
-
-            try:
-                rslt = fmin_bfgs(
-                    opt_obj.crit_func,
-                    x_optim_free_scaled_start,
-                    args=args,
-                    gtol=bfgs_gtol,
-                    epsilon=bfgs_eps,
-                    maxiter=bfgs_maxiter,
-                    full_output=True,
-                    disp=False,
-                )
-
-                success = rslt[6] not in [1, 2]
-                message = "Optimization terminated successfully."
-                if rslt[6] == 1:
-                    message = "Maximum number of iterations exceeded."
-                elif rslt[6] == 2:
-                    message = "Gradient and/or function calls not changing."
-
-            except MaxfunError:
-                success = False
-                message = "Maximum number of iterations exceeded."
-
-        elif optimizer_used == "SCIPY-LBFGSB":
-
-            lbfgsb_maxiter = optimizer_options["SCIPY-LBFGSB"]["maxiter"]
-            lbfgsb_maxls = optimizer_options["SCIPY-LBFGSB"]["maxls"]
-            lbfgsb_factr = optimizer_options["SCIPY-LBFGSB"]["factr"]
-            lbfgsb_pgtol = optimizer_options["SCIPY-LBFGSB"]["pgtol"]
-            lbfgsb_eps = optimizer_options["SCIPY-LBFGSB"]["eps"]
-            lbfgsb_m = optimizer_options["SCIPY-LBFGSB"]["m"]
-
-            try:
-                rslt = fmin_l_bfgs_b(
-                    opt_obj.crit_func,
-                    x_optim_free_scaled_start,
-                    args=args,
-                    approx_grad=True,
-                    bounds=paras_bounds_free_scaled,
-                    m=lbfgsb_m,
-                    factr=lbfgsb_factr,
-                    pgtol=lbfgsb_pgtol,
-                    epsilon=lbfgsb_eps,
-                    iprint=-1,
-                    maxfun=maxfun,
-                    maxiter=lbfgsb_maxiter,
-                    maxls=lbfgsb_maxls,
-                )
-
-                success = rslt[2]["warnflag"] in [0]
-                message = rslt[2]["task"]
-
-            except MaxfunError:
-                success = False
-                message = "Maximum number of iterations exceeded."
-
-        elif optimizer_used == "SCIPY-POWELL":
-
-            powell_maxiter = optimizer_options["SCIPY-POWELL"]["maxiter"]
-            powell_maxfun = optimizer_options["SCIPY-POWELL"]["maxfun"]
-            powell_xtol = optimizer_options["SCIPY-POWELL"]["xtol"]
-            powell_ftol = optimizer_options["SCIPY-POWELL"]["ftol"]
-
-            try:
-                rslt = fmin_powell(
-                    opt_obj.crit_func,
-                    x_optim_free_scaled_start,
-                    args,
-                    powell_xtol,
-                    powell_ftol,
-                    powell_maxiter,
-                    powell_maxfun,
-                    disp=0,
-                )
-
-                success = rslt[5] not in [1, 2]
-                message = "Optimization terminated successfully."
-                if rslt[5] == 1:
-                    message = "Maximum number of function evaluations."
-                elif rslt[5] == 2:
-                    message = "Maximum number of iterations."
-
-            except MaxfunError:
-                success = False
-                message = "Maximum number of iterations exceeded."
-
-        else:
-            raise NotImplementedError
-
-        record_estimation_final(success, message)
-        record_estimation_stop()
-
-    elif request == "simulate":
-
-        # Draw draws for the simulation.
-        periods_draws_sims = create_draws(
-            num_periods, num_agents_sim, seed_sim, is_debug
-        )
-
-        # Draw standard normal deviates for the solution and evaluation step.
-        periods_draws_emax = create_draws(
-            num_periods, num_draws_emax, seed_emax, is_debug
-        )
-
-        # Collect arguments for different implementations of the simulation.
-        state_space = pyth_solve(
-            interpolation,
-            num_points_interp,
-            num_periods,
-            periods_draws_emax,
-            edu_spec,
-            optim_paras,
-            num_types,
-        )
-
-        simulated_data = pyth_simulate(
-            state_space, num_agents_sim, periods_draws_sims, edu_spec, optim_paras
-        )
-
-        args = (state_space, simulated_data)
-
-    else:
-        raise NotImplementedError("This request is not implemented.")
-
-    return args
-
-
-def get_precondition_matrix(
-    precond_spec,
-    optim_paras,
-    x_optim_all_unscaled_start,
-    args,
-    maxfun,
-    num_paras,
-    num_types,
-):
-    """ Get the preconditioning matrix for the optimization.
-    """
-    # Auxiliary objects
-    num_free = optim_paras["paras_fixed"].count(False)
-
-    # Set up a special instance of the optimization class.
-    precond_matrix = np.identity(num_free)
-
-    opt_obj = OptimizationClass(
-        x_optim_all_unscaled_start,
-        optim_paras["paras_fixed"],
-        precond_matrix,
-        num_types,
-    )
-
-    opt_obj.is_scaling = False
-
-    # Distribute information about user request.
-    precond_minimum = precond_spec["minimum"]
-    precond_type = precond_spec["type"]
-    precond_eps = precond_spec["eps"]
-
-    # Get the subset of free parameters for subsequent numerical
-    # approximation of the gradient.
-    x_optim_free_unscaled_start = []
-    for i in range(num_paras):
-        if not optim_paras["paras_fixed"][i]:
-            x_optim_free_unscaled_start += [x_optim_all_unscaled_start[i]]
-
-    if precond_type == "identity" or maxfun == 0:
-        precond_matrix = np.identity(num_free)
-    elif precond_type == "magnitudes":
-        precond_matrix = get_scales_magnitudes(x_optim_free_unscaled_start)
-    elif precond_type == "gradient":
-        opt_obj.is_scaling = None
-        grad = approx_fprime(
-            x_optim_free_unscaled_start, opt_obj.crit_func, precond_eps, *args
-        )
-        precond_matrix = np.zeros((num_free, num_free))
-        for i in range(num_free):
-            grad[i] = max(np.abs(grad[i]), precond_minimum)
-            precond_matrix[i, i] = grad[i]
-    else:
-        raise NotImplementedError
-
-    # Write out scaling matrix to allow for restart.
-    np.savetxt("scaling.respy.out", precond_matrix, fmt="%45.15f")
-
-    return precond_matrix
-
-
-def get_scales_magnitudes(x_optim_free_unscaled_start):
-    """Calculate scaling factor based on the magnitudes of the variables."""
-    # Auxiliary objects
-    num_free = len(x_optim_free_unscaled_start)
-
-    # Initialize container
-    precond_matrix = np.zeros((num_free, num_free))
-
-    for i, x in enumerate(x_optim_free_unscaled_start):
-        # Special case
-        if x == 0.0:
-            scale = 1
-        else:
-            magnitude = int(floor(log10(abs(x))))
-            if magnitude == 0:
-                scale = 1.0 / 10.0
-            else:
-                scale = (10 ** magnitude) ** (-1) / 10.0
-        precond_matrix[i, i] = scale
-
-    return precond_matrix
-
-
-def minimal_simulation_interface(attr):
-    # Draw draws for the simulation.
-    periods_draws_sims = create_draws(
-        attr["num_periods"], attr["num_agents_sim"], attr["seed_sim"], attr["is_debug"]
-    )
-
-    # Draw standard normal deviates for the solution and evaluation step.
-    periods_draws_emax = create_draws(
-        attr["num_periods"], attr["num_draws_emax"], attr["seed_emax"], attr["is_debug"]
-    )
-
-    state_space = pyth_solve(
-        attr["interpolation"],
-        attr["num_points_interp"],
-        attr["num_periods"],
-        periods_draws_emax,
-        attr["edu_spec"],
-        attr["optim_paras"],
-        attr["num_types"],
-    )
-
-    simulated_data = pyth_simulate(
-        state_space,
-        attr["num_agents_sim"],
-        periods_draws_sims,
-        attr["edu_spec"],
-        attr["optim_paras"],
-    )
-
-    return state_space, simulated_data
-
-
-def minimal_estimation_interface(attr, df):
-    def get_independent_estimation_arguments(attr):
-        periods_draws_prob = create_draws(
-            attr["num_periods"],
-            attr["num_draws_prob"],
-            attr["seed_prob"],
-            attr["is_debug"],
-        )
-        periods_draws_emax = create_draws(
-            attr["num_periods"],
-            attr["num_draws_emax"],
-            attr["seed_emax"],
-            attr["is_debug"],
-        )
-
-        # Construct the state space
-        state_space = StateSpace(
-            attr["num_periods"],
-            attr["num_types"],
-            attr["edu_spec"]["start"],
-            attr["edu_spec"]["max"],
-        )
-
-        return periods_draws_emax, periods_draws_prob, state_space
-
-    def get_parameter_vector(attr):
-        """This function is obsolete with estimagic."""
-        x = get_optim_paras(
-            attr["optim_paras"], attr["num_paras"], "all", attr["is_debug"]
-        )
-
-        return x
-
-    def process_dataset(attr, df):
-        """Process the dataset from disk."""
-
-        # Process dataset from files.
-        df.set_index(["Identifier", "Period"], drop=False, inplace=True)
-
-        # We want to allow to estimate with only a subset of periods in the sample.
-        cond = df["Period"] < attr["num_periods"]
-        df = df[cond]
-
-        # Only keep the information that is relevant for the estimation.
-        # Once that is done,  impose some type restrictions.
-        df = df[DATA_LABELS_EST]
-        df = df.astype(DATA_FORMATS_EST)
-
-        # We want to restrict the sample to meet the specified initial conditions.
-        cond = df["Years_Schooling"].loc[:, 0].isin(attr["edu_spec"]["start"])
-        df.set_index(["Identifier"], drop=False, inplace=True)
-        df = df.loc[cond]
-
-        # We now subset the dataframe to include only the number of agents that are
-        # requested for the estimation. However, this requires to adjust the
-        # num_agents_est as the dataset might actually be smaller as we restrict
-        # initial conditions.
-        df = df.loc[df.index.unique()[: attr["num_agents_est"]]]
-        df.set_index(["Identifier", "Period"], drop=False, inplace=True)
-
-        # We need to update the number of individuals for the estimation as the
-        # whole dataset might actually be lower.
-        num_agents_est = df["Identifier"].nunique()
-
-        return df, num_agents_est
-
-    # Process data
-    df, attr["num_agents_est"] = process_dataset(attr, df)
-
-    # Collect arguments for estimation.
-    (
-        periods_draws_emax,
-        periods_draws_prob,
-        state_space,
-    ) = get_independent_estimation_arguments(attr)
-    x = get_parameter_vector(attr)
-
-    log_likelihood = partial(
-        pyth_criterion,
-        interpolation=attr["interpolation"],
-        num_points_interp=attr["num_points_interp"],
-        is_debug=attr["is_debug"],
-        data=df,
-        tau=attr["tau"],
-        periods_draws_emax=periods_draws_emax,
-        periods_draws_prob=periods_draws_prob,
-        state_space=state_space,
-    )
-
-    return x, log_likelihood(x)
