+++ setup.py
@@ -37,13 +37,7 @@ setup(
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
     ],
-    install_requires=[
-        "numba>=0.43",
-        "pandas>=0.24",
-        "scipy>=0.19",
-        "pytest>=4.0",
-        "pyaml",
-    ],
+    install_requires=["numba>=0.43", "pandas>=0.24", "pytest>=4.0", "pyaml"],
     platforms="any",
     include_package_data=True,
     zip_safe=False,
