+++ tox.ini
@@ -15,7 +15,6 @@ conda_deps =
     pytest
     pytest-cov
     pytest-xdist
-    scipy >= 1.2.1
 conda_channels =
     numba
     defaults
@@ -57,9 +56,10 @@ max-line-length = 89
 [flake8]
 max-line-length = 88
 ignore =
-    D  ; ignores docstring style errors, enable if you are nit-picky
-    E203  ; ignores whitespace around : which is enforced by Black
-    W503  ; ignores linebreak before binary operator which is enforced by Black
+    D  ; ignores docstring style errors, enable if you are nit-picky.
+    E203  ; ignores whitespace around : which is enforced by Black.
+    W503  ; ignores linebreak before binary operator which is enforced by Black.
+    T000  ; ignores TODOs.
 
 [pytest]
 addopts = --doctest-modules -n auto
